EDSC-4162 Refactoring some files and adding feature toggle for tour
