+++ static.config.json
@@ -51,7 +51,8 @@
     "disableSwodlr": "false",
     "macOSEddDownloadSize":130,
     "windowsEddDownloadSize":100,
-    "linuxEddDownloadSize":90
+    "linuxEddDownloadSize":90,
+    "disableSiteTour": "false"
   },
   "environment": {
     "test": {
